Add flavors of demurrage token to dev deployment

See merge request grassrootseconomics/cic-internal-integration!181

+++ apps/util/requirements/update_base.sh
@@ -9,16 +9,18 @@ fi
 t=$(mktemp)
 >&2 echo using tmp $t
 
-repos=(../../cic-cache ../../cic-eth ../../cic-ussd ../../data-seeding ../../cic-notify)
+repos=(../../cic-cache ../../cic-eth ../../cic-ussd ../../data-seeding ../../cic-notify ../../contract-migration)
 
 for r in ${repos[@]}; do
 	f="$r/requirements.txt"
 	>&2 echo updating $f
-	pyreq-update $f base_requirement.txt -vv > $t 
+	pyreq-update $f base_requirement.txt -v > $t 
 	cp $t $f
 
 	f="$r/test_requirements.txt"
-	>&2 echo updating $f
-	pyreq-update $f base_requirement.txt -vv > $t 
-	cp $t $f
+	if [ -f $f ]; then
+		>&2 echo updating $f
+		pyreq-update $f base_requirement.txt -v > $t
+		cp $t $f
+	fi
 done
